// src/app/shared/models/setting.model.ts
export class SettingDto {
    constructor() {
    }
    stId: string | null;
    stOldVersion: string | null;
    stVersionApp: string | null;
    stCode: number | null;
    stMessage: string | null;
    stLink: string | null;
    stActive: number | null;
    stCheckAtMobile: number | null;
    stCreatedDate: Date | null;
    stUpdatedDate: Date | null;
    stCreatedBy: string | null;
  }
  